Move IndexBackfiller from 'local' to 'index' and fix some related, and unrelated, build errors
