+++ packages/firestore/test/unit/index/index_backfiller.test.ts
@@ -19,7 +19,7 @@ import { expect } from 'chai';
 
 import { User } from '../../../src/auth/user';
 import { Query, queryToTarget } from '../../../src/core/query';
-import { IndexBackfiller } from '../../../src/local/index_backfiller';
+import { IndexBackfiller } from '../../../src/index/index_backfiller';
 import { IndexedDbPersistence } from '../../../src/local/indexeddb_persistence';
 import { LocalStore } from '../../../src/local/local_store';
 import { newLocalStore } from '../../../src/local/local_store_impl';
@@ -41,11 +41,11 @@ import { newAsyncQueue } from '../../../src/util/async_queue_impl';
 import { key, version } from '../../util/helpers';
 import * as Helpers from '../../util/helpers';
 
-import { CountingQueryEngine } from './counting_query_engine';
-import { JSON_SERIALIZER } from './persistence_test_helpers';
-import * as PersistenceTestHelpers from './persistence_test_helpers';
-import { TestDocumentOverlayCache } from './test_document_overlay_cache';
-import { TestIndexManager } from './test_index_manager';
+import { CountingQueryEngine } from '../local/counting_query_engine';
+import { JSON_SERIALIZER } from '../local/persistence_test_helpers';
+import * as PersistenceTestHelpers from '../local/persistence_test_helpers';
+import { TestDocumentOverlayCache } from '../local/test_document_overlay_cache';
+import { TestIndexManager } from '../local/test_index_manager';
 
 describe('IndexedDb IndexBackfiller', () => {
   if (!IndexedDbPersistence.isAvailable()) {
