+++ packages/firestore/test/unit/local/counting_query_engine.ts
@@ -81,17 +81,14 @@ export class CountingQueryEngine extends QueryEngine {
     );
   }
 
-  initialize(
-    localDocuments: LocalDocumentsView,
-    indexManager: IndexManager
-  ): void {
+  initialize(localDocuments: LocalDocumentsView): void {
     const view = new LocalDocumentsView(
       this.wrapRemoteDocumentCache(localDocuments.remoteDocumentCache),
       localDocuments.mutationQueue,
       this.wrapOverlayCache(localDocuments.documentOverlayCache),
       localDocuments.indexManager
     );
-    return super.initialize(view, indexManager);
+    return super.initialize(view);
   }
 
   private wrapRemoteDocumentCache(
